fix: align Authority id type with database schema

+++ src/main/java/kg/attractor/jobsearch/repository/AuthorityRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AuthorityRepository extends JpaRepository<Authority, Long> {
+public interface AuthorityRepository extends JpaRepository<Authority, Integer> {
     Optional<Authority> findByAuthority(String authority);
 }
